export interfaces for client method options
