+++ packages/seal/src/client.ts
@@ -24,34 +24,18 @@ import {
 	fetchKeysForAllIds,
 } from './key-server.js';
 import type { DerivedKey, KeyServer } from './key-server.js';
-import type { SessionKey } from './session-key.js';
-import type { KeyCacheKey, SealCompatibleClient } from './types.js';
+import type {
+	DecryptOptions,
+	EncryptOptions,
+	FetchKeysOptions,
+	GetDerivedKeysOptions,
+	KeyCacheKey,
+	KeyServerConfig,
+	SealClientExtensionOptions,
+	SealClientOptions,
+	SealCompatibleClient,
+} from './types.js';
 import { createFullId, count } from './utils.js';
-
-/**
- * Configuration options for initializing a SealClient
- * @property serverConfigs: Array of key server configs consisting of objectId, weight, optional API key name and API key.
- * @property verifyKeyServers: Whether to verify the key servers' authenticity.
- * 	 Should be false if servers are pre-verified (e.g., getAllowlistedKeyServers).
- * 	 Defaults to true.
- * @property timeout: Timeout in milliseconds for network requests. Defaults to 10 seconds.
- */
-export interface SealClientExtensionOptions {
-	serverConfigs: KeyServerConfig[];
-	verifyKeyServers?: boolean;
-	timeout?: number;
-}
-
-export interface KeyServerConfig {
-	objectId: string;
-	weight: number;
-	apiKeyName?: string;
-	apiKey?: string;
-}
-
-export interface SealClientOptions extends SealClientExtensionOptions {
-	suiClient: SealCompatibleClient;
-}
 
 export class SealClient {
 	#suiClient: SealCompatibleClient;
@@ -122,15 +106,7 @@ export class SealClient {
 		id,
 		data,
 		aad = new Uint8Array(),
-	}: {
-		kemType?: KemType;
-		demType?: DemType;
-		threshold: number;
-		packageId: string;
-		id: string;
-		data: Uint8Array;
-		aad?: Uint8Array;
-	}) {
+	}: EncryptOptions) {
 		const packageObj = await this.#suiClient.core.getObject({ objectId: packageId });
 		if (String(packageObj.object.version) !== '1') {
 			throw new InvalidPackageError(`Package ${packageId} is not the first version`);
@@ -166,15 +142,7 @@ export class SealClient {
 	 * @param txBytes - The transaction bytes to use (that calls seal_approve* functions).
 	 * @returns - The decrypted plaintext corresponding to ciphertext.
 	 */
-	async decrypt({
-		data,
-		sessionKey,
-		txBytes,
-	}: {
-		data: Uint8Array;
-		sessionKey: SessionKey;
-		txBytes: Uint8Array;
-	}) {
+	async decrypt({ data, sessionKey, txBytes }: DecryptOptions) {
 		const encryptedObject = EncryptedObject.parse(data);
 
 		this.#validateEncryptionServices(
@@ -277,17 +245,7 @@ export class SealClient {
 	 * @param sessionKey - The session key to use.
 	 * @param threshold - The threshold for the TSS encryptions. The function returns when a threshold of key servers had returned keys for all ids.
 	 */
-	async fetchKeys({
-		ids,
-		txBytes,
-		sessionKey,
-		threshold,
-	}: {
-		ids: string[];
-		txBytes: Uint8Array;
-		sessionKey: SessionKey;
-		threshold: number;
-	}) {
+	async fetchKeys({ ids, txBytes, sessionKey, threshold }: FetchKeysOptions) {
 		if (threshold > this.#totalWeight || threshold < 1) {
 			throw new InvalidThresholdError(
 				`Invalid threshold ${threshold} servers with weights ${this.#configs}`,
@@ -399,13 +357,7 @@ export class SealClient {
 		txBytes,
 		sessionKey,
 		threshold,
-	}: {
-		kemType?: KemType;
-		id: string;
-		txBytes: Uint8Array;
-		sessionKey: SessionKey;
-		threshold: number;
-	}): Promise<Map<string, DerivedKey>> {
+	}: GetDerivedKeysOptions): Promise<Map<string, DerivedKey>> {
 		switch (kemType) {
 			case KemType.BonehFranklinBLS12381DemCCA:
 				const keyServers = await this.getKeyServers();
