+++ packages/seal/src/index.ts
@@ -3,7 +3,16 @@
 
 export { getAllowlistedKeyServers } from './key-server.js';
 export { EncryptedObject } from './bcs.js';
-export { SealClient, type SealClientOptions, type KeyServerConfig } from './client.js';
+export { SealClient } from './client.js';
 export { SessionKey, type ExportedSessionKey } from './session-key.js';
 export * from './error.js';
-export type { SealCompatibleClient } from './types.js';
+export type {
+	SealCompatibleClient,
+	SealClientOptions,
+	SealClientExtensionOptions,
+	KeyServerConfig,
+	EncryptOptions,
+	DecryptOptions,
+	FetchKeysOptions,
+	GetDerivedKeysOptions,
+} from './types.js';
